+++ tests/domain_test.py
@@ -1,4 +1,5 @@
 from isotools import Transcriptome
+from isotools.domains import add_domains_to_table
 
 
 def test_anno_domains():
@@ -8,3 +9,6 @@ def test_anno_domains():
     ref_tr = next(tr for tr in g.ref_transcripts if tr['transcript_name'] == 'FN1-207')
     dom = ref_tr['domain']['annotation']
     assert len(dom) == 25, 'expected 25 annotation domains in FN1-207, but found {len(dom)}'
+
+    diffexpr = isoseq.altsplice_test(groups=isoseq.groups()).sort_values('pvalue')
+    diffexpr = add_domains_to_table(diffexpr, isoseq, 'annotation', insert_after='nmdB')